+++ Sources/Network/Endpoints/AccountEndpoint.swift
@@ -18,53 +18,29 @@ extension Endpoints {
 }
 
 public struct Account: Decodable {
-    public let state: State
-    public let credentials: Credential?
-    public let validationURL: URL?
-    public let message: String?
-    public let mandate: Mandate?
+    public let credentials: [Credentials]
+    public let validationURL: URL
 
-    public enum State: String, Decodable {
-        case pending = "PENDING"
-        case successful = "SUCCESSFUL"
-        case failed = "FAILED"
-        case error = "ERROR"
-    }
+    let urlScheme: String
 
     enum CodingKeys: String, CodingKey {
-        case id
-        case state
         case credentials
         case validationURL = "validationLink"
-        case message
-        case mandate
     }
 
     public init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
-        self.state = try container.decode(Account.State.self, forKey: .state)
-        switch state {
-        case .pending:
-            self.validationURL = try container.decode(URL.self, forKey: .validationURL)
-            self.credentials = nil
-            self.message = nil
-            self.mandate = nil
-        case .successful:
-            self.validationURL = nil
-            self.credentials = try container.decode(Credential.self, forKey: .credentials)
-            self.message = nil
-#warning("Remove optional try if backend has added it")
-            self.mandate = try? container.decode(Mandate.self, forKey: .mandate)
-        case .error, .failed:
-            self.validationURL = nil
-            self.credentials = nil
-            self.message = try container.decode(String.self, forKey: .message)
-            self.mandate = nil
+        self.credentials = try container.decode([Account.Credentials].self, forKey: .credentials)
+        self.validationURL = try container.decode(URL.self, forKey: .validationURL)
+
+        guard let urlScheme = decoder.userInfo[.urlScheme] as? String else {
+            throw DecodingError.valueNotFound(String.self, .init(codingPath: [], debugDescription: "Missing URLScheme in decoders userInfo"))
         }
+        self.urlScheme = urlScheme
     }
 
-    public static func validateCallbackURL(_ url: URL, forScheme scheme: String) -> Bool {
-        guard url.scheme == scheme,
+    public func validateCallbackURL(_ url: URL) -> Bool {
+        guard url.scheme == urlScheme,
               url.host == "account",
               url.lastPathComponent == "validation" else {
             return false
@@ -75,29 +51,6 @@ public struct Account: Decodable {
 
 extension Account: Equatable {
     public static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.state == rhs.state &&
-        lhs.credentials == rhs.credentials &&
-        lhs.validationURL == rhs.validationURL &&
-        lhs.message == rhs.message
-    }
-}
-
-public struct Credential: Decodable {
-    public let id: ID
-    public let name: String
-    public let holderName: String
-    public let currencyCode: CurrencyCode
-    public let bank: String
-    public let createdAt: Date
-    public let iban: IBAN
-
-    public typealias ID = Tagged<Credential, String>
-    public typealias IBAN = Tagged<(Credential, iban: ()), String>
-    public typealias CurrencyCode = Tagged<(Credential, currencyCode: ()), String>
-}
-
-extension Credential: Equatable {
-    public static func == (lhs: Self, rhs: Self) -> Bool {
-        lhs.id == rhs.id
+        lhs.credentials == rhs.credentials
     }
 }