+++ Sources/Network/Endpoints/RegisterEndpoint.swift
@@ -24,12 +24,15 @@ extension Endpoints {
 public struct App: Codable {
     public let identifier: Idenitifer
     public let secret: Secret
+    public let urlScheme: UrlScheme
 
     public typealias Idenitifer = Tagged<(App, identifier: ()), String>
     public typealias Secret = Tagged<(App, secret: ()), String>
+    public typealias UrlScheme = Tagged<(App, urlScheme: ()), String>
 
     enum CodingKeys: String, CodingKey {
         case identifier = "appIdentifier"
         case secret = "appSecret"
+        case urlScheme = "appUrlScheme"
     }
 }