+++ Tests/Network/Endpoints/RegisterEndpointTests.swift
@@ -8,7 +8,7 @@
 import XCTest
 @testable import SnabblePayNetwork
 
-final class CredentialsEndpointTests: XCTestCase {
+final class RegisterEndpointTests: XCTestCase {
     func testEndpoint() throws {
         let endpoint = Endpoints.Register.post(customUrlScheme: "snabble-pay", apiKeyValue: "123456")
         XCTAssertEqual(endpoint.path, "/apps/register")
@@ -34,5 +34,6 @@ final class CredentialsEndpointTests: XCTestCase {
         let app = try JSONDecoder().decode(App.self, from: registerData)
         XCTAssertEqual(app.identifier, "1l2z79uvnKU18hJ621hDti2Q1mckTs8633HFlUz7PCG1OalckFyKf/TzJlGcOUC4WPInc+RrKCAPLc0loJCtRw==")
         XCTAssertEqual(app.secret, "qPgwvqkVCFn+aFxljTClV7+kTe+18rOQ7Qrdp5YSethhi2X9Sp97UiDkAO3qzXgcdDi/+VazutfHxbA4SZKYWA==")
+        XCTAssertEqual(app.urlScheme, "snabble-pay")
     }
 }